requirements.txt
spacy
pdfplumber==0.11.0
python-docx==1.1.0